[LibOS] Use the syscall interface for custom calls
This change modifies Gramine's custom calls (`register_library`,
`run_test`) to use the same interface as system calls (with a dummy
syscall number). This is so that these calls are executed on LibOS's
stack, not the user application stack, which will make it possible to
implement ASan stack sanitization.

At the same time, names in the syscall interface are updated
(`gramine_entry_api.h`, `GRAMINE_SYSCALL` etc.).

This change requires rebuilding the patched Glibc. To force Meson to do
that automatically, we rename the subproject from `glibc-2.34` to
`glibc-2.34-1`.

Signed-off-by: Paweł Marczewski <pawel@invisiblethingslab.com>